Refactor payment context classes and update PaymentSessionsClient authorization type
